refactoring types & blob storage, adding better image support
