+++ package.json
@@ -74,6 +74,7 @@
     "react-hook-form": "^7.51.3",
     "react-markdown": "^9.0.1",
     "remark-gfm": "^4.0.0",
+    "sharp": "^0.33.3",
     "sonner": "^1.4.41",
     "string-strip-html": "^8.4.0",
     "tailwind-merge": "^2.2.1",
