+++ src/main/lib/images.ts
@@ -1,0 +1,1 @@
+import { config } from "@shared/config";
