+++ src/renderer/src/components/PersonaFormModal.tsx
@@ -3,7 +3,7 @@ import { Input } from "@/components/ui/input";
 import { Textarea } from "@/components/ui/textarea";
 import { PencilSquareIcon } from "@heroicons/react/24/solid";
 import { zodResolver } from "@hookform/resolvers/zod";
-import { PersonaFormData, personaFormSchema } from "@shared/types";
+import { PersonaFormData, personaFormSchema } from "@shared/forms";
 import { useEffect, useRef, useState } from "react";
 import { Controller, useForm } from "react-hook-form";
 import { toast } from "sonner";
