+++ src/shared/types.ts
@@ -66,107 +66,43 @@ export const cardSchema = z.object({
 });
 export type CardData = z.infer<typeof cardSchema>;
 
-// Forms
-// ===========================================
-export const personaFormSchema = z.object({
-  name: z
-    .string()
-    .min(config.persona.nameMinChars)
-    .max(config.persona.nameMaxChars)
-    .regex(/^[a-zA-Z0-9 -]*$/, "Name can only contain letters, numbers, spaces, and hyphens"),
-  description: z.string().max(config.persona.descriptionMaxChars),
-  isDefault: z.boolean(),
-  avatarURI: z.string().optional(),
-  // TODO: implement persona banners
-  bannerURI: z.string().optional()
-});
-export type PersonaFormData = z.infer<typeof personaFormSchema>;
-
-// Card Form Schema
-// Used for validating the form data before creating a card
-// Some fields are more lenient, some extra fields added (e.g. avatarURI, bannerURI)
-const characterFormSchema = z.object({
-  name: z
-    .string()
-    .min(config.card.nameMinChars)
-    .max(config.card.nameMaxChars)
-    .regex(
-      /^[\p{L}\p{N}_ -]+$/u,
-      "Invalid input. Only Unicode letters, numbers, spaces, hyphens, and underscores are allowed."
-    ),
-  handle: z
-    .string()
-    .min(config.card.handleMinChars)
-    .max(config.card.handleMaxChars)
-    .regex(/^[a-zA-Z0-9_-]*$/, "Handle can only contain letters, numbers, and dashes")
-    .optional(),
-
-  description: z.string().min(config.card.descriptionMinChars).max(config.card.descriptionMaxChars),
-  greeting: z.string().min(config.card.greetingMinChars).max(config.card.greetingMaxChars),
-  msg_examples: z.string().min(config.card.msgExamplesMinChars).max(config.card.msgExamplesMaxChars),
-  avatarURI: z.string().optional(),
-  bannerURI: z.string().optional()
-});
-
-const worldFormSchema = z.object({
-  description: z.string().min(config.card.descriptionMinChars).max(config.card.descriptionMaxChars)
-});
-const metaFormSchema = z.object({
-  title: z
-    .string()
-    .min(config.card.titleMinChars)
-    .max(config.card.titleMaxChars)
-    .regex(
-      /^[\p{L}\p{N}_ -]+$/u,
-      "Invalid input. Only Unicode letters, numbers, spaces, hyphens, and underscores are allowed."
-    ),
-  notes: z.string().min(config.card.notesMinChars).max(config.card.notesMaxChars),
-  tagline: z.string().min(config.card.tagLineMinChars).max(config.card.taglineMaxChars),
-  // TODO: this should be z.array(z.string()) instead, this is hacky
-  // Change this when you polish tag support
-  // Use React Aria https://react-spectrum.adobe.com/react-aria/TagGroup.html
-  // Tags are a list of strings, all lowercase, alphanumeric, with dashes and spaces allowed.
-  tags: z
-    .string()
-    .min(1)
-    .max(512)
-    .regex(/^[a-z0-9-\s]+$/, "Tag must be lowercase alphanumeric with dash and spaces allowed.")
-});
-
-export const cardFormSchema = z.object({
-  character: characterFormSchema,
-  world: worldFormSchema,
-  meta: metaFormSchema
-});
-
-// Contents of the card's directory
-// =====================================
-export interface CardBundle {
-  id: number;
+// =========================================================
+//  Card Storage
+// =========================================================
+// Contents of the card directory
+export type CardBundle = {
   data: CardData;
   avatarURI: string;
   bannerURI: string;
+};
+
+export interface UICardBundle extends CardBundle {
+  id: string;
 }
 
-export interface CardBundleWithoutID {
+export interface BackendCardBundle {
   data: CardData;
-  avatarURI: string;
-  bannerURI: string;
+  avatarFilePath: string;
+  bannerFilePath: string;
 }
 
-// Persona
-// =====================================
+// =========================================================
+//  Persona Storage
+// =========================================================
+// Contents of the persona directory
+export interface PersonaBundle {
+  avatarURI: string;
+}
+
 export interface PersonaData extends Persona {}
-// Contents of the persona's directory
-export interface PersonaBundleWithoutData {
-  avatarURI: string;
-}
-export interface PersonaBundle extends PersonaBundleWithoutData {
+
+export interface UIPersonaBundle extends PersonaBundle {
   data: PersonaData;
 }
 
-// Settings from settings.json
-// =====================================
+// =========================================================
+//  Settings Store
+// =========================================================
 export interface Settings {
   chat: {
     provider: ProviderE;
@@ -185,9 +121,6 @@ export interface Settings {
   };
 }
 
-export type CardFormData = z.infer<typeof cardFormSchema>;
-
 export type Result<T, E> = { kind: "ok"; value: T } | { kind: "err"; error: E };
-
 export type ImageExt = (typeof supportedImageExts)[number];
 export type supportedCardExts = (typeof supportedImageExts)[number];
